+++ adm_program/modules/downloads/get_file.php
@@ -15,7 +15,7 @@
 require('../../system/common.php');
 
 // Initialize and check the parameters
-$getFileId = admFuncVariableIsValid($_GET, 'file_id', 'int', array('requireValue' => true));
+$getFileId = admFuncVariableIsValid($_GET, 'file_id', 'numeric', array('requireValue' => true));
 
 // pruefen ob das Modul ueberhaupt aktiviert ist
 if ($gPreferences['enable_download_module'] != 1)
