+++ adm_program/modules/members/members_function.php
@@ -26,8 +26,8 @@ require_once('../../system/common.php');
 require_once('../../system/login_valid.php');
 
 // Initialize and check the parameters
-$getUserId = admFuncVariableIsValid($_GET, 'usr_id', 'int', array('requireValue' => true));
-$getMode   = admFuncVariableIsValid($_GET, 'mode',   'int', array('requireValue' => true));
+$getUserId = admFuncVariableIsValid($_GET, 'usr_id', 'numeric', array('requireValue' => true));
+$getMode   = admFuncVariableIsValid($_GET, 'mode',   'numeric', array('requireValue' => true));
 
 // nur berechtigte User duerfen Funktionen aufrufen
 if(!$gCurrentUser->editUsers())
