+++ adm_program/modules/photos/photo_album_function.php
@@ -19,7 +19,7 @@ require_once('../../system/common.php');
 require_once('../../system/login_valid.php');
 
 // Initialize and check the parameters
-$getPhotoId = admFuncVariableIsValid($_GET, 'pho_id', 'int');
+$getPhotoId = admFuncVariableIsValid($_GET, 'pho_id', 'numeric');
 $getMode    = admFuncVariableIsValid($_GET, 'mode',   'string', array('requireValue' => true, 'validValues' => array('new', 'change', 'delete')));
 
 // pruefen ob das Modul ueberhaupt aktiviert ist
