+++ adm_program/modules/rooms/rooms_function.php
@@ -17,8 +17,8 @@
 require_once('../../system/common.php');
 
 // Initialize and check the parameters
-$getRoomId = admFuncVariableIsValid($_GET, 'room_id', 'int');
-$getMode   = admFuncVariableIsValid($_GET, 'mode',    'int', array('requireValue' => true));
+$getRoomId = admFuncVariableIsValid($_GET, 'room_id', 'numeric');
+$getMode   = admFuncVariableIsValid($_GET, 'mode',    'numeric', array('requireValue' => true));
 
 // nur berechtigte User duerfen die Profilfelder bearbeiten
 if (!$gCurrentUser->isWebmaster())
