+++ adm_program/modules/rooms/rooms_new.php
@@ -18,7 +18,7 @@ require_once('../../system/common.php');
 require_once('../../system/login_valid.php');
 
 // Initialize and check the parameters
-$getRoomId   = admFuncVariableIsValid($_GET, 'room_id',  'int');
+$getRoomId   = admFuncVariableIsValid($_GET, 'room_id',  'numeric');
 $getHeadline = admFuncVariableIsValid($_GET, 'headline', 'string', array('defaultValue' => $gL10n->get('SYS_ROOM')));
 
 // nur berechtigte User duerfen die Profilfelder bearbeiten
